woah mama did a whole bunch more stuff added lift recording too sleepy to add odo reset its like four lines but ill write it tomorrow also cleaned everything up and simplified it a lot so it will (hopefully?) be less taxing computationally and also more readable i also added parking and i also created a controller diagram on google slides

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Mechanisms/Lift.java
@@ -49,7 +49,7 @@ public class Lift {
     public int getSlidePosition() { // returns integer average of the two slides
         return (transferM1.getCurrentPosition() + transferM2.getCurrentPosition()) / 2;
     }
-    public void movePID(int desiredHeight) { //Likely only going to be used for auto
+    public void movePID(double desiredHeight) { //Likely only going to be used for auto
         pidController = new PID(desiredHeight - getSlidePosition(), p, i, d); //resets data values
         while(desiredHeight != getSlidePosition() || desiredHeight != getSlidePosition()){
             pidController.update(desiredHeight-getSlidePosition());
